+++ packages/exports/src/dataHelpers/prepLessonPlanForDocs.ts
@@ -1,7 +1,8 @@
+import { keyStageToSentenceCase } from "@oakai/core/src/data/parseKeyStage";
+
 import type { LessonPlanDocInputData } from "../schema/input.schema";
 import type { LessonPlanDocsTemplateData } from "../schema/lessonPlanDocsTemplate.schema";
 import {
-  camelCaseToTitleCase,
   formatLegacyQuestionForDoc,
   formatQuestionForDoc,
   stringOrBullets,
@@ -22,7 +23,7 @@ export async function prepLessonPlanForDocs(
   return Promise.resolve({
     lesson_title: lessonPlan.title,
     subject: lessonPlan.subject,
-    key_stage: camelCaseToTitleCase(lessonPlan.keyStage) ?? "",
+    key_stage: keyStageToSentenceCase(lessonPlan.keyStage ?? ""),
     topic: lessonPlan.topic ?? " ",
     learning_outcome: lessonPlan.learningOutcome ?? " ",
     learning_cycle_outcome_1: lessonPlan.learningCycles?.[0]